requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.9
outcome==1.2.0
Pillow==10.0.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
werkzeug==2.3.7
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
